recast data to float32

+++ segmentation_code/Patient.py
@@ -80,7 +80,7 @@ class Patient:
 
             imgs[idx-1] = pixel_data
 
-        return imgs.astype(np.float16)
+        return imgs.astype(np.float32)
